# pkg/stringid: optimize GenerateRandomID